Extract dial gesture hook
